Driving School Management Software/requirements.txt
fpdf==2.5.7
Pillow==10.0.0
